feat: Initialise node package, use prettier and package-lint, and add self workflows (#105)
first i wanted to prettier the workflows, then i realised i needed a
package.json

then if you have a package.json you might as well lint it

and then i thought about releasing, and isn't it weird that we don't
release this like all the other projects ...

so i made "self" workflows that applie to this project.

they use the shared workflows, of course.

it's a bit weird naming. it would be nicer if all the other workflows
were named "shared" or moved to a shared folder or something, but then
we'd have to update EVERY PROJECT AT THE SAME TIME ...

so it's kinda just easier to name them "self" i reckon.

i'm excitied about the auto-update latest because that's always a faff
to do manually. hope it works!